fix(module8-task1): Добавил хук usePageInit
